Pass ID to peripherals

+++ include/peripherals/nrf52832/rtc.h
@@ -4,6 +4,6 @@
 
 #define RTC_MAX_CC 4
 
-PERIPHERAL(RTC, rtc, size_t cc_num, cpu_t **cpu, uint32_t id)
+PERIPHERAL(RTC, rtc, size_t cc_num, cpu_t **cpu, uint8_t id)
 
 void rtc_tick(RTC_t *rtc);
